proto_tls: new module tls_mgm
