+++ modules/tls_mgm/tls_config.h
@@ -35,18 +35,7 @@
 #ifndef tls_config_h
 #define tls_config_h
 
-enum tls_method {
-	TLS_METHOD_UNSPEC = 0,
-	TLS_USE_TLSv1_cli,
-	TLS_USE_TLSv1_srv,
-	TLS_USE_TLSv1,
-	TLS_USE_SSLv23_cli,
-	TLS_USE_SSLv23_srv,
-	TLS_USE_SSLv23,
-	TLS_USE_TLSv1_2_cli,
-	TLS_USE_TLSv1_2_srv,
-	TLS_USE_TLSv1_2
-};
+#include "tls_config_helper.h"
 
 extern int      tls_method;
